+++ language/go/BUILD.bazel
@@ -75,7 +75,6 @@ go_test(
         "//config:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/testtools:go_default_library",
-        "//internal/walk:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//language/proto:go_default_library",
@@ -84,5 +83,6 @@ go_test(
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
+        "//walk:go_default_library",
     ],
 )
