+++ language/proto/BUILD.bazel
@@ -49,13 +49,13 @@ go_test(
         "//config:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/testtools:go_default_library",
-        "//internal/walk:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
+        "//walk:go_default_library",
     ],
 )
